+++ src/features/output/elements/Button/returnInput.js
@@ -1,4 +1,4 @@
-import animateElement from './animateElement'
+import animateElement from '../../../shared/utils/animateElement'
 import outputAnimations from '../../components/Output.animations.json'
 import changeClass from '../../../../utils/changeClass'
 import providerAnimations from '../../../input/components/Provider.animations.json'
